Factor out getACS and getTransactionTrees
https://github.com/digital-asset/daml/pull/8898#discussion_r579146541

changelog_begin
changelog_end
